refactor(cli): move next-work materialize behind mto adapter
